fea: create user DAO
- DAO means Data Access Object
