+++ src/app/app/layout.tsx
@@ -5,7 +5,7 @@ import { authOptions } from "@/shared/auth";
 import Navbar from "@/components/Navbar";
 import AuthProvider from "@/contexts/AuthProvider";
 import { Container } from "@/shared/components";
-import { fetchUserWithSession } from "@/shared/db";
+import { fetchUserWithSession } from "@/shared/auth";
 
 export const metadata: Metadata = {
   title: "PodCodar",
